+++ src/components/Question/index.jsx
@@ -2,13 +2,13 @@ import React, { useState, useEffect } from 'react'
 import PropTypes from 'prop-types'
 import './styles.css'
 
+import StarsRating from '../StarsRating'
+import { Button } from 'react-bootstrap'
+
 import { shuffleAnswers } from '../../utils'
 import { ACTION_TYPES } from '../../reducers/quiz.reducer'
 
-import StarsRating from '../StarsRating'
-import { Container, Row, Col, Button } from 'react-bootstrap'
-
-function Question({ dispatch, question, isAnswerSelected, isAnswerCorrect }) {
+function Question ({ dispatch, question, isAnswerSelected, isAnswerCorrect }) {
   // QUESTION CHOICES
 
   const [questionChoices, setQuestionChoices] = useState([])
@@ -19,12 +19,14 @@ function Question({ dispatch, question, isAnswerSelected, isAnswerCorrect }) {
     )
   }, [question])
 
+  // SELECT ANSWER HANDLER
+
   const [selectedButtonIndex, setSelectedButtonIndex] = useState(null)
 
   const handleSelectAnswer = (value, index) => {
     setSelectedButtonIndex(index)
 
-    const isAnswerCorrect = value === question.correct_answer ? true : false
+    const isAnswerCorrect = value === question.correct_answer
 
     dispatch({ type: ACTION_TYPES.ANSWER_SELECTED, payload: isAnswerCorrect })
     dispatch({ type: ACTION_TYPES.UPDATE_SCORE })
@@ -54,26 +56,19 @@ function Question({ dispatch, question, isAnswerSelected, isAnswerCorrect }) {
 
       <h5 className='question-text'>{decodeURIComponent(question.question)}</h5>
 
-      <Container>
-        <Row xs={1} sm={2}>
-          {questionChoices.map((value, index) => (
-            <Col
-              key={index}
-              className='answer__container'
-              sm={{ span: 5, offset: index % 2 !== 0 ? 2 : 0 }}
-            >
-              <Button
-                block
-                disabled={isAnswerSelected}
-                variant={getButtonVariant(value, index)}
-                onClick={() => handleSelectAnswer(value, index)}
-              >
-                {decodeURIComponent(value)}
-              </Button>
-            </Col>
-          ))}
-        </Row>
-      </Container>
+      <div className='answers__container'>
+        {questionChoices.map((value, index) => (
+          <Button
+            key={value}
+            className='answer-button'
+            disabled={isAnswerSelected}
+            variant={getButtonVariant(value, index)}
+            onClick={() => handleSelectAnswer(value, index)}
+          >
+            {decodeURIComponent(value)}
+          </Button>
+        ))}
+      </div>
     </div>
   )
 }