+++ src/components/ScoreProgress/index.jsx
@@ -3,15 +3,19 @@ import './styles.css'
 import PropTypes from 'prop-types'
 import { ProgressBar } from 'react-bootstrap'
 
-function ScoreProgress({ minScore, currentScore, maxScore }) {
+function ScoreProgress ({ minScore, currentScore, maxScore }) {
   return (
     <>
       <div className='score-progress__status'>
-        <p>Score: {currentScore.toFixed(0)}%</p>
-        <p>Max Score: {maxScore.toFixed(0)}%</p>
+        <p>Score: {Math.round(currentScore)}%</p>
+        <p>Max Score: {Math.round(maxScore)}%</p>
       </div>
+
       <ProgressBar className='score-progress__bar'>
-        <ProgressBar variant='danger' now={minScore} />
+        <ProgressBar
+          variant={minScore === maxScore ? 'success' : 'danger'}
+          now={minScore}
+        />
         <ProgressBar variant='success' now={currentScore - minScore} />
         <ProgressBar variant='warning' now={maxScore - currentScore} />
       </ProgressBar>