+++ src/components/StarsRating/index.jsx
@@ -2,17 +2,13 @@ import React, { useCallback } from 'react'
 import PropTypes from 'prop-types'
 import StarRatingComponent from 'react-star-rating-component'
 
-function StarsRating({ difficulty }) {
+function StarsRating ({ difficulty }) {
   const getDifficultyValue = useCallback(() => {
     switch (difficulty) {
-      case 'hard':
-        return 3
-      case 'medium':
-        return 2
-      case 'easy':
-        return 1
-      default:
-        return 0
+      case 'hard': return 3
+      case 'medium': return 2
+      case 'easy': return 1
+      default: return 0
     }
   }, [difficulty])